+++ mslib/msui/tableview.py
@@ -36,13 +36,12 @@ import types
 
 from mslib.msui import hexagon_dockwidget as hex
 from mslib.msui import performance_settings as perfset
-from PyQt5 import QtWidgets, QtGui, QtCore
+from PyQt5 import QtWidgets, QtGui
 from mslib.msui.qt5 import ui_tableview_window as ui
 from mslib.utils.qt import dropEvent, dragEnterEvent
 from mslib.msui import flighttrack as ft
 from mslib.msui.viewwindows import MSUIViewWindow
 from mslib.msui.icons import icons
-from mslib.utils.config import save_settings_qsettings
 
 try:
     import mpl_toolkits.basemap.pyproj as pyproj
@@ -91,24 +90,6 @@ class MSUITableViewWindow(MSUIViewWindow, ui.Ui_TableViewWindow):
         self.cbTools.currentIndexChanged.connect(self.openTool)
 
         self.resizeColumns()
-
-    def changeEvent(self, event):
-        top_left = self.mapToGlobal(QtCore.QPoint(0, 0))
-        if top_left.x() != 0:
-            os_screen_region = [top_left.x(), top_left.y(), self.width(), self.height()]
-            settings = {'os_screen_region': os_screen_region}
-            # we have to save this to reuse it by the tutorials
-            save_settings_qsettings(self.settings_tag, settings)
-        QtWidgets.QWidget.changeEvent(self, event)
-
-    def moveEvent(self, event):
-        top_left = self.mapToGlobal(QtCore.QPoint(0, 0))
-        if top_left.x() != 0:
-            os_screen_region = [top_left.x(), top_left.y(), self.width(), self.height()]
-            settings = {'os_screen_region': os_screen_region}
-            # we have to save this to reuse it by the tutorials
-            save_settings_qsettings(self.settings_tag, settings)
-        QtWidgets.QWidget.moveEvent(self, event)
 
     def setPerformance(self, settings):
         """
