+++ mslib/msui/topview.py
@@ -30,9 +30,8 @@
 
 import functools
 import logging
-import pyautogui as pag
 
-from mslib.utils.config import config_loader, save_settings_qsettings
+from mslib.utils.config import config_loader
 from mslib.utils.coordinate import get_projection_params
 from PyQt5 import QtGui, QtWidgets, QtCore
 from mslib.msui.qt5 import ui_topview_window as ui
@@ -259,28 +258,6 @@ class MSUITopViewWindow(MSUIMplViewWindow, ui.Ui_TopViewWindow):
 
     def __del__(self):
         del self.mpl.canvas.waypoints_interactor
-
-    def changeEvent(self, event):
-        top_left = self.mapToGlobal(QtCore.QPoint(0, 0))
-        if top_left.x() != 0:
-            os_screen_region = [top_left.x(), top_left.y(), self.width(), self.height()]
-            settings = {'os_screen_region': os_screen_region}
-            # we have to save this to reuse it by the tutorials
-            save_settings_qsettings(self.settings_tag, settings)
-            im = pag.screenshot(region=os_screen_region)
-            im.save('/tmp/msui/tv_changing.png')
-        QtWidgets.QWidget.changeEvent(self, event)
-
-    def moveEvent(self, event):
-        top_left = self.mapToGlobal(QtCore.QPoint(0, 0))
-        if top_left.x() != 0:
-            os_screen_region = [top_left.x(), top_left.y(), self.width(), self.height()]
-            settings = {'os_screen_region': os_screen_region}
-            # we have to save this to reuse it by the tutorials
-            save_settings_qsettings(self.settings_tag, settings)
-            im = pag.screenshot(region=os_screen_region)
-            im.save('/tmp/msui/tv_moveing.png')
-        QtWidgets.QWidget.moveEvent(self, event)
 
     @QtCore.pyqtSlot(ft.WaypointsTableModel)
     def update_active_flighttrack(self, active_flighttrack):
